// src/configs/global.enum.ts
export enum AddCart {
  ADD_TO_CART_CALLER = 'ADD_TO_CART_CALLER',
  ADD_TO_CART = 'ADD_TO_CART',
}

export enum PaymentSuccess {
  PAYMENT_CALLER = 'PAYMENT_CALLER',
  PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
}

export enum GetFeeds {
  GET_FEEDS_CALLER = 'GET_FEEDS_CALLER',
  GET_FEEDS_DEFAULT = 'GET_FEEDS_DEFAULT',
  GET_FEEDS_STARTED = 'GET_FEEDS_STARTED',
  GET_FEEDS_FAILED = 'GET_FEEDS_FAILED',
  GET_FEEDS_SUCCESS = 'GET_FEEDS_SUCCESS',
}

export enum GetUserPosts {
  GET_USER_POSTS_CALLER = 'GET_USER_POSTS_CALLER',
  GET_USER_POSTS_DEFAULT = 'GET_USER_POSTS_DEFAULT',
  GET_USER_POSTS_STARTED = 'GET_USER_POSTS_STARTED',
  GET_USER_POSTS_FAILED = 'GET_USER_POSTS_FAILED',
  GET_USER_POSTS_SUCCESS = 'GET_USER_POSTS_SUCCESS',
}

export enum Post {
  POST_CALLER = 'POST_CALLER',
  POST_DEFAULT = 'POST_DEFAULT',
  POST_STARTED = 'POST_STARTED',
  POST_FAILED = 'POST_FAILED',
  POST_SUCCESS = 'POST_SUCCESS',
}

export enum GetCategories {
  GET_CATEGORIES_CALLER = 'GET_CATEGORIES_CALLER',
  GET_CATEGORIES_DEFAULT = 'GET_CATEGORIES_DEFAULT',
  GET_CATEGORIES_STARTED = 'GET_CATEGORIES_STARTED',
  GET_CATEGORIES_FAILED = 'GET_CATEGORIES_FAILED',
  GET_CATEGORIES_SUCCESS = 'GET_CATEGORIES_SUCCESS',
}